third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/w3c/webperf/submission/Intel/resource-timing/test_resource_timing_cross_origin_redirect_chain_allow_timing-expected.txt

This is a testharness.js-based test.
PASS window.performance is defined 
PASS Starting document.location.hostname is correct (127.0.0.1:8000) 
PASS redirectStart should be greater than 0 in timing allowed cross-origin redirect!  
PASS redirectStart should be equal to startTime in timing allowed cross-origin redirect! 
PASS redirectEnd should be 3 secs later than redirectStart 
PASS fetchStart should be no less than redirectEnd in timing allowed cross-origin redirect! 
Harness: the test ran to completion.
